Separate test framework tests
